robot/requirements.txt
pytesseract~=0.3.10
opencv-contrib-python==4.5.5.62
numpy
bestconfig
pywebview